backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
sqlmodel
authlib
pyjwt
httpx
itsdangerous
apscheduler
